+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Version 0.0.3 (development)
+
+- Exit with error code if log file does not exist
+- Set `--azureml` automatically when running on Azure ML
+
 ## Version 0.0.2 (development)
 
 - Created package: marian-tensorboard